+++ packages/threadplane-middleware/src/threadplane/middleware/langgraph/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: MIT
-"""threadplane-client-tools — LangGraph middleware for client-declared tools."""
+"""threadplane-middleware — LangGraph middleware for client-declared tools."""
 
-from threadplane.client_tools.middleware import (
+from threadplane.middleware.langgraph.middleware import (
     bind_client_tools,
     client_tool_names,
     client_tool_specs,
